myblog/requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
billiard==3.5.0.5
celery==4.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-ckeditor==5.7.1
django-haystack==2.8.1
django-js-asset==1.2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-redis-sessions==0.5.6
django-simple-captcha==0.5.12
django-taggit==1.1.0
django-tinymce==2.6.0
dnspython==1.16.0
eventlet==0.24.1
fdfs-client-py==1.2.6
future==0.17.1
greenlet==0.4.15
httplib2==0.13.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jieba==0.39
joblib==0.13.2
jsonfield==2.0.2
kiwisolver==1.1.0
kombu==4.5.0
lxml==4.3.3
Markdown==3.1.1
matplotlib==3.1.0
monotonic==1.5
mutagen==1.42.0
mysqlclient==1.4.4
numpy==1.16.3
oauthlib==3.1.0
pandas==0.24.2
parsel==1.5.1
Pillow==5.4.1
playsound==1.2.2
prettytable==0.7.2
pure-sasl==0.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodomex==3.7.2
PyDispatcher==2.0.5
PyExecJS==1.5.1
pygame==1.9.6
Pygments==2.4.2
PyHamcrest==1.9.0
PyHive==0.6.1
pymongo==3.8.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-alipay-sdk==1.10.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
pywin32==224
queuelib==1.5.0
recordtype==1.3
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
sasl==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
Scrapy==1.6.0
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
thrift==0.11.0
thrift-sasl==0.3.0
Twisted==19.2.0
urllib3==1.24.3
urlparse2==1.1.1
vine==1.3.0
w3lib==1.20.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
